* origin/sherlock:
  MB-15021 Change the return type of getMemorySize() to int64_t.

+++ utils/get_memory_size.cc
@@ -21,7 +21,7 @@
 /**
  * Returns the size of physical memory (RAM) in bytes.
  */
-size_t get_memory_size(void)
+int64_t get_memory_size(void)
 {
 #if defined(_WIN32) && (defined(__CYGWIN__) || defined(__CYGWIN32__))
     /* Cygwin under Windows. ------------------------------------ */
@@ -29,7 +29,7 @@ size_t get_memory_size(void)
     MEMORYSTATUS status;
     status.dwLength = sizeof(status);
     GlobalMemoryStatus( &status );
-    return (size_t)status.dwTotalPhys;
+    return (int64_t)status.dwTotalPhys;
 
 #elif defined(WIN32) || defined(_WIN32)
     /* Windows. ------------------------------------------------- */
@@ -37,7 +37,7 @@ size_t get_memory_size(void)
     MEMORYSTATUSEX status;
     status.dwLength = sizeof(status);
     GlobalMemoryStatusEx( &status );
-    return (size_t)status.ullTotalPhys;
+    return (int64_t)status.ullTotalPhys;
 
 #else
     /* UNIX variants. ------------------------------------------- */
@@ -52,23 +52,23 @@ size_t get_memory_size(void)
     mib[1] = HW_PHYSMEM64; /* NetBSD, OpenBSD. --------- */
 #endif
     int64_t size = 0; /* 64-bit */
-    size_t len = sizeof( size );
+    size_t len = sizeof(size);
     if ( sysctl( mib, 2, &size, &len, NULL, 0 ) == 0 ) {
-        return (size_t)size;
+        return size;
     }
-    return 0L; /* Failed? */
+    return 0; /* Failed? */
 
 #elif defined(_SC_AIX_REALMEM)
     /* AIX. ----------------------------------------------------- */
-    return (size_t)sysconf( _SC_AIX_REALMEM ) * (size_t)1024L;
+    return (int64_t)(sysconf( _SC_AIX_REALMEM ) * 1024);
 
 #elif defined(_SC_PHYS_PAGES) && defined(_SC_PAGESIZE)
     /* FreeBSD, Linux, OpenBSD, and Solaris. -------------------- */
-    return (size_t)sysconf( _SC_PHYS_PAGES ) * (size_t)sysconf( _SC_PAGESIZE );
+    return (int64_t)(sysconf( _SC_PHYS_PAGES ) * sysconf( _SC_PAGESIZE ));
 
 #elif defined(_SC_PHYS_PAGES) && defined(_SC_PAGE_SIZE)
     /* Legacy. -------------------------------------------------- */
-    return (size_t)sysconf( _SC_PHYS_PAGES ) * (size_t)sysconf( _SC_PAGE_SIZE );
+    return (int64_t)(sysconf( _SC_PHYS_PAGES ) * sysconf( _SC_PAGE_SIZE ));
 
 #elif defined(CTL_HW) && (defined(HW_PHYSMEM) || defined(HW_REALMEM))
     /* DragonFly BSD, FreeBSD, NetBSD, OpenBSD, and OSX. -------- */
@@ -82,9 +82,9 @@ size_t get_memory_size(void)
     unsigned int size = 0; /* 32-bit */
     size_t len = sizeof( size );
     if ( sysctl( mib, 2, &size, &len, NULL, 0 ) == 0 ) {
-        return (size_t)size;
+        return (int64_t) size;
     }
-    return 0L; /* Failed? */
+    return 0; /* Failed? */
 #endif /* sysctl and sysconf variants */
 
 #endif
